Move PREFIX to DaybreakPlugin
